#27 Move importImage components
